requirements.txt
streamlit==1.32.0
opencv-python==4.9.0.80
plotly==5.20.0
deepface==0.0.91
Pillow==10.2.0
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
scipy==1.12.0
tf-keras